pyproject.toml
[project]
name = "lumina"
version = "0.2.4"
description = "Lumina: Modernized Gemini Image Generation CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
    "pillow>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "types-pillow>=10.2.0",
]

[project.scripts]
lumina = "lumina.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/lumina"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
check_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
